bot/requirements.txt
discord-py-interactions
aiohttp
toml
redis>=4.2.0rc1